+++ unreal/Plugins/RaftSim/Source/RaftSimEditor/Private/Commands/RaftSimEditorEnvironmentAutomation.cpp
@@ -50,7 +50,9 @@ bool FRaftSimEditorModule::TickPhotorealEnvironmentAutomationStartup(float)
 
     if (bCreateLandscapeImportCandidateMapsOnStartup)
     {
-        bSucceeded &= CreateLandscapeImportCandidateMaps(Summary);
+        bSucceeded &= CreateLandscapeImportCandidateMaps(
+            Summary,
+            LandscapeImportCandidateRiverFilter);
     }
 
     if (bCreatePhotorealRiverWaterMaterialOnStartup)